+++ src/Labs/ElevensLab/Activity9StarterCode/Board.java
@@ -119,7 +119,7 @@ public abstract class Board {
 	 */
 	public void replaceSelectedCards(List<Integer> selectedCards) {
 		for (Integer k : selectedCards) {
-			deal(k.intValue());
+			deal(k);
 		}
 	}
